Move deprecations log into a tab of the error log page, check permission in processor
