+++ manager/assets/modext/widgets/system/modx.grid.deprecated.log.js
@@ -70,11 +70,41 @@ MODx.grid.DeprecatedLogGrid = function(config) {
             ,groupTextTpl: '{group:htmlEncode}<br>' +
                 '<tpl if="values.rs[0].data.method_since"><span style="font-weight: normal;">{[_("deprecated_since")]}</span> v{[Ext.util.Format.htmlEncode(values.rs[0].data.method_since)]} | </tpl>' +
                 '<span style="font-weight: normal;">{[Ext.util.Format.htmlEncode(values.rs[0].data.method_recommendation)]}</span>'
-        })
+        }),
+
+        tbar: [{
+            text: _('clear')
+            ,id: 'modx-abtn-clear'
+            ,cls: 'primary-button'
+            ,handler: this.clearDeprecatedLog
+            ,scope: this
+            ,hidden: !MODx.perm.error_log_erase
+        }]
     });
 
     MODx.grid.DeprecatedLogGrid.superclass.constructor.call(this,config);
 };
 Ext.extend(MODx.grid.DeprecatedLogGrid, MODx.grid.Grid, {
+    clearDeprecatedLog: function() {
+        var panel = Ext.getCmp('modx-panel-deprecated-log');
+        panel.el.mask(_('working'));
+        MODx.Ajax.request({
+            url: this.config.url
+            ,params: {
+                action: 'System/DeprecatedLog/Clear'
+            }
+            ,listeners: {
+                'success': {fn:function(r) {
+                        panel.el.unmask();
+                        MODx.msg.status({
+                            title: _('success'),
+                            message: _('deprecated_log_cleared')
+                        });
+                        // refresh grid
+                        this.refresh();
+                    },scope: this}
+            }
+        });
+    }
 });
 Ext.reg('modx-grid-deprecated-log', MODx.grid.DeprecatedLogGrid);
