+++ manager/controllers/default/system/event.class.php
@@ -28,6 +28,8 @@ class SystemEventManagerController extends modManagerController {
     public function loadCustomCssJs() {
         $mgrUrl = $this->modx->getOption('manager_url',null,MODX_MANAGER_URL);
         $this->addJavascript($mgrUrl.'assets/modext/widgets/system/modx.panel.error.log.js');
+        $this->addJavascript($mgrUrl.'assets/modext/widgets/system/modx.grid.deprecated.log.js');
+        $this->addJavascript($mgrUrl.'assets/modext/widgets/system/modx.panel.deprecated.log.js');
         $this->addJavascript($mgrUrl.'assets/modext/sections/system/error.log.js');
         $this->addHtml('<script>
         MODx.hasEraseErrorLog = "'.($this->modx->hasPermission('error_log_erase') ? 1 : 0).'"
